style: working on viewmodel
