+++ frontend/EatAndTell/app/src/main/java/com/example/eatandtell/di/RetrofitClient.kt
@@ -1,0 +1,17 @@
+import com.squareup.moshi.Moshi
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+
+object RetrofitClient {
+
+    private const val BaseURL= "http://ec2-13-125-91-166.ap-northeast-2.compute.amazonaws.com/users/"
+
+    private val moshi = Moshi.Builder().build()
+
+    val retrofit = {
+        Retrofit.Builder()
+            .baseUrl(BaseURL) // actual backend URL
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .build()
+    }
+}
