+++ frontend/EatAndTell/app/src/main/java/com/example/eatandtell/dto/Register.kt
@@ -1,15 +1,16 @@
 package com.example.eatandtell.dto
 
-import retrofit2.Response
-
 data class RegisterRequest(
     val username: String,
     val password: String,
     val email: String,
 )
 
-data class RegisterResponse(val token: String)
-
-data class RegisterResult(
-    val token: String,
+data class RegisterResponse(
+    val token: String
 )
+
+sealed class RegistrationResult {
+    data class Success(val token: String?) : RegistrationResult()
+    data class Error(val errorMessage: String) : RegistrationResult()
+}
