+++ frontend/EatAndTell/app/src/main/java/com/example/eatandtell/ui/login/LoginViewModel.kt
@@ -1,25 +1,22 @@
 package com.example.eatandtell.ui.login
+import RetrofitClient
+import RetrofitClient.retrofit
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.example.eatandtell.ApiService
+import com.example.eatandtell.dto.LoginRequest
+import com.example.eatandtell.dto.LoginResponse
+import com.example.eatandtell.dto.LoginResult
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
 import retrofit2.Call
+import retrofit2.Retrofit
 import retrofit2.Callback
 import retrofit2.http.Body
 import retrofit2.http.POST
 import retrofit2.Response
-import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-
-data class LoginRequest(val username: String, val password: String)
-data class LoginResponse(val token: String)
-
-interface ApiService {
-    @POST("login/") // The login endpoint (hypothetical)
-    fun loginUser(@Body loginData: LoginRequest): Call<LoginResponse>
-}
-
 
 class LoginViewModel : ViewModel() {
     //interface for callback
@@ -31,12 +28,7 @@ class LoginViewModel : ViewModel() {
     private val _loginResult = MutableLiveData<LoginResult>()
     val loginResult: LiveData<LoginResult> = _loginResult
 
-    val retrofit = Retrofit.Builder()
-        .baseUrl("http://ec2-13-125-91-166.ap-northeast-2.compute.amazonaws.com/users/") // actual backend URL
-        .addConverterFactory(MoshiConverterFactory.create()) // Add coroutine adapter
-        .build()
-
-    private val apiService = retrofit.create(ApiService::class.java)
+    private val apiService = RetrofitClient.retrofit.create(ApiService::class.java)
     fun loginUser(username: String, password: String,  callback: LoginCallback) {
         val loginData = LoginRequest(username, password)
         val call = apiService.loginUser(loginData)
@@ -63,11 +55,6 @@ class LoginViewModel : ViewModel() {
         })
     }
 
-    //Sealed Class (closer to mvvm)
-    sealed class LoginResult {
-        data class Success(val token: String?) : LoginResult()
-        data class Error(val errorMessage: String) : LoginResult()
-    }
 
 }
