+++ frontend/EatAndTell/app/src/main/java/com/example/eatandtell/ui/signup/RegisterActivity.kt
@@ -28,10 +28,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.eatandtell.Logo
 import com.example.eatandtell.R
+import com.example.eatandtell.ui.CustomTextField
+import com.example.eatandtell.ui.Logo
 import com.example.eatandtell.ui.login.LoginActivity
 import com.example.eatandtell.ui.login.LoginScreen
+import com.example.eatandtell.ui.showToast
 
 class RegisterActivity : ComponentActivity() {
     private val registerViewModel: RegisterViewModel by viewModels()
@@ -71,6 +73,8 @@ fun SignupScreen(context: ComponentActivity) {
     }
 
     var passwordHidden by rememberSaveable { mutableStateOf(true) }
+
+    //TODO: var들은 remember 안해도 button event로만 트리거하면 될 것 같은데? + CustomTextField 재사용하기
 
     // Main content of SignupActivity
     Column(
@@ -243,8 +247,6 @@ fun SignupButton (
                 }
             }
 
-
-
         },
         colors = ButtonDefaults.buttonColors(
             containerColor = Color(0xFFF23F18),
@@ -259,43 +261,8 @@ fun SignupButton (
     }
 }
 
-private fun showToast(context: Context, message: String) {
-    Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-    println(message)
-}
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CustomTextField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    placeholder: String,
-    supportingText: String = "",
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-    trailingIcon: (@Composable () -> Unit)? = null,
-    modifier: Modifier = Modifier
-) {
-    TextField(
-        value = value,
-        onValueChange = onValueChange,
-        visualTransformation = visualTransformation,
-        trailingIcon = trailingIcon,
-        modifier = modifier,
-        colors = TextFieldDefaults.textFieldColors(
-            containerColor = Color(0xFFEEEEEE),
-            cursorColor = Color.Black,
-            focusedIndicatorColor = Color(0xFFA0A0A0),
-            unfocusedIndicatorColor = Color.Transparent,
 
-            ),
-        placeholder = { Text(placeholder, style = TextStyle(
-            fontSize = 12.sp,
-            fontWeight = FontWeight(400),
-            color = Color(0xFF000000)
-        )) },
-        maxLines = 1
-    )
-}
 
 @Composable
 fun PasswordVisibilityToggle(passwordHidden: Boolean, onClick: () -> Unit) {
