+++ frontend/EatAndTell/app/src/main/java/com/example/eatandtell/ui/signup/RegisterViewModel.kt
@@ -1,7 +1,13 @@
 package com.example.eatandtell.ui.signup
+import RetrofitClient
+import RetrofitClient.retrofit
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.example.eatandtell.ApiService
+import com.example.eatandtell.dto.RegisterRequest
+import com.example.eatandtell.dto.RegisterResponse
+import com.example.eatandtell.dto.RegistrationResult
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
 import retrofit2.Call
 import retrofit2.Callback
@@ -10,17 +16,6 @@ import retrofit2.http.POST
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-
-
-data class RegisterRequest(val username: String, val password: String, val email:String)
-data class RegisterResponse(val token: String)
-
-interface ApiService {
-    @POST("register/") // The registration endpoint
-    fun registerUser(@Body registrationData: RegisterRequest): Call<RegisterResponse>
-
-}
-
 
 class RegisterViewModel : ViewModel() {
     //interface for callback
@@ -31,13 +26,8 @@ class RegisterViewModel : ViewModel() {
 
     private val _registerResult = MutableLiveData<RegistrationResult>()
     val registerResult: LiveData<RegistrationResult> = _registerResult
-    val BaseURL= "http://ec2-13-125-91-166.ap-northeast-2.compute.amazonaws.com/users/"
-    val retrofit = Retrofit.Builder()
-        .baseUrl(BaseURL) // actual backend URL
-        .addConverterFactory(MoshiConverterFactory.create())
-        .build()
 
-    private val apiService = retrofit.create(ApiService::class.java)
+    private val apiService = RetrofitClient.retrofit.create(ApiService::class.java)
 
     fun registerUser(username: String, password: String, email: String, callback: RegisterCallback) {
         val registrationData = RegisterRequest(username, password, email)
@@ -63,12 +53,6 @@ class RegisterViewModel : ViewModel() {
                 callback.onRegisterError(errorMessage)
             }
         })
-    }
-
-    //Sealed Class (closer to mvvm)
-    sealed class RegistrationResult {
-        data class Success(val token: String?) : RegistrationResult()
-        data class Error(val errorMessage: String) : RegistrationResult()
     }
 
 }
